fixed label_tensor float conversion

+++ scripts/old_dataset_loader.py
@@ -83,7 +83,6 @@ class MicroscopyDataset(Dataset):
         label_tensor = torch.tensor(label, dtype=torch.float)
 
         if self.label == 'classification':
-            label_tensor = 1 if label_tensor >= 25 else 0
-        label_tensor = label_tensor.float()
+            label_tensor = float(1 if label_tensor >= 25 else 0)
 
         return combined_image, label_tensor
